+++ jdaviz/configs/cubeviz/plugins/viewers.py
@@ -2,12 +2,13 @@ from glue.core import BaseData
 from glue_jupyter.bqplot.image import BqplotImageView
 
 from jdaviz.core.registries import viewer_registry
+from jdaviz.configs.default.plugins.viewers import JdavizViewerMixin
 
 __all__ = ['CubevizImageView']
 
 
 @viewer_registry("cubeviz-image-viewer", label="Image 2D (Cubeviz)")
-class CubevizImageView(BqplotImageView):
+class CubevizImageView(BqplotImageView, JdavizViewerMixin):
     # Whether to inherit tools from glue-jupyter automatically. Set this to
     # False to have full control here over which tools are shown in case new
     # ones are added in glue-jupyter in future that we don't want here.