+++ jdaviz/configs/default/plugins/viewers.py
@@ -4,8 +4,20 @@ from glue_jupyter.table import TableViewer
 
 from jdaviz.core.registries import viewer_registry
 
-__all__ = []
+__all__ = ['JdavizViewerMixin']
 
 viewer_registry.add("g-profile-viewer", label="Profile 1D", cls=BqplotProfileView)
 viewer_registry.add("g-image-viewer", label="Image 2D", cls=BqplotImageView)
 viewer_registry.add("g-table-viewer", label="Table", cls=TableViewer)
+
+
+class JdavizViewerMixin:
+    @property
+    def jdaviz_app(self):
+        """The Jdaviz application tied to the viewer."""
+        return self.session.jdaviz_app
+
+    @property
+    def jdaviz_helper(self):
+        """The Jdaviz configuration helper tied to the viewer."""
+        return self.jdaviz_app._jdaviz_helper